refactor: made tasks a top level module in core.
