+++ cli/src/commands/enrich/epss.rs
@@ -4,7 +4,7 @@ use std::sync::Arc;
 use crate::commands::enrich::EnrichArgs;
 use harbcore::entities::tasks::{Task, TaskKind};
 use harbcore::services::enrichments::vulnerabilities::epss::EpssScoreTask;
-use harbcore::services::tasks::TaskProvider;
+use harbcore::tasks::TaskProvider;
 use platform::persistence::mongodb::{Context, Store};
 
 use crate::Error;
