+++ cli/src/commands/enrich/snyk.rs
@@ -10,7 +10,7 @@ use harbcore::services::enrichments::vulnerabilities::{
 };
 use harbcore::services::packages::PackageService;
 use harbcore::services::snyk::SnykService;
-use harbcore::services::tasks::TaskProvider;
+use harbcore::tasks::TaskProvider;
 use platform::persistence::mongodb::{Context, Store};
 
 use crate::Error;
