+++ sdk/core/src/services/enrichments/vulnerabilities/epss/mod.rs
@@ -2,7 +2,7 @@ mod client;
 
 use crate::entities::enrichments::Vulnerability;
 use crate::entities::tasks::Task;
-use crate::services::tasks::TaskProvider;
+use crate::tasks::TaskProvider;
 use crate::Error;
 use async_trait::async_trait;
 use client::Client;
