+++ sdk/core/src/services/enrichments/vulnerabilities/snyk/mod.rs
@@ -11,7 +11,7 @@ mod tests {
     };
     use crate::services::packages::PackageService;
     use crate::services::snyk::SnykService;
-    use crate::services::tasks::TaskProvider;
+    use crate::tasks::TaskProvider;
     use crate::{entities, Error};
     use platform::persistence::mongodb::Store;
     use std::sync::Arc;
