+++ sdk/core/src/services/enrichments/vulnerabilities/snyk/provider.rs
@@ -5,7 +5,7 @@ use crate::entities::xrefs::XrefKind;
 use crate::services::enrichments::vulnerabilities::VulnerabilityService;
 use crate::services::packages::PackageService;
 use crate::services::snyk::{SnykService, SNYK_DISCRIMINATOR};
-use crate::services::tasks::TaskProvider;
+use crate::tasks::TaskProvider;
 use crate::Error;
 use async_trait::async_trait;
 use platform::persistence::mongodb::{Service, Store};
