+++ sdk/core/src/services/mod.rs
@@ -30,8 +30,5 @@ pub mod teams;
 /// cross-references for entities that implement the [Xrefs] trait.
 pub mod xrefs;
 
-/// The [tasks] module contains domain logic and traits related to processing tasks.
-pub mod tasks;
-
 /// The [analytics] module contains reporting services to make the data useful.
 pub mod analytics;
