+++ sdk/extension-template/src/commands/example.rs
@@ -1,6 +1,6 @@
 use clap::Parser;
 use harbcore::entities::tasks::{Task, TaskKind};
-use harbcore::services::tasks::TaskProvider;
+use harbcore::tasks::TaskProvider;
 
 use crate::tasks::example::ExampleTask;
 use crate::Error;
